Reverted back to simplier structure #10
Hopefully retained all the changes that we want to keep. Namely:
*** Previous values are now being displayed when the pop-up opens.
This works now for New Inspection Date as well as all the other window.prompt
calls.

Officially calls the result.cancel() when the user hits the cancel button
of the DatePickerDialog. Without this, the Javascript side can freeze.

To determine whether a window.prompt call is going to be a DatePickerDialog,
I'm now searching just if the message contains the string "date" rather than
if it ends with it.

Made button labels into proper R.string strings.

** Next Inspection Date dialog is now a date picker on the Android #945.